+++ __test__/tools/create-configs.ts
@@ -1,7 +1,7 @@
 import { BundlibPkgJson, configsFromPkg } from "../../src/api";
 
 async function createConfigs(cwd: string, dev: boolean, pkgJson?: BundlibPkgJson) {
-  return pkgJson ? configsFromPkg(cwd, dev, false, pkgJson) : configsFromPkg(cwd, dev);
+  return pkgJson ? configsFromPkg(cwd, null, pkgJson) : configsFromPkg(cwd, { dev });
 }
 
 export default createConfigs;